pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-policy-engine"
version = "1.0.0"
description = "A multi-objective, event-driven governance simulation environment for reinforcement learning agents. OpenEnv-compatible."
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "AI Policy Engine Team"}]
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.25.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["ai_policy_engine", "ai_policy_engine.server"]
package-dir = { "ai_policy_engine" = ".", "ai_policy_engine.server" = "server" }
